Factor code common to signals_{byt,nat}.c and move it to signals.c
Then, signals_byt.c becomes empty and can be removed.
