--- runtime/signals_byt.c
@@ -1,77 +1,0 @@
-/**************************************************************************/
-/*                                                                        */
-/*                                 OCaml                                  */
-/*                                                                        */
-/*          Xavier Leroy and Damien Doligez, INRIA Rocquencourt           */
-/*                                                                        */
-/*   Copyright 2007 Institut National de Recherche en Informatique et     */
-/*     en Automatique.                                                    */
-/*                                                                        */
-/*   All rights reserved.  This file is distributed under the terms of    */
-/*   the GNU Lesser General Public License version 2.1, with the          */
-/*   special exception on linking described in the file LICENSE.          */
-/*                                                                        */
-/**************************************************************************/
-
-#define CAML_INTERNALS
-
-/* Signal handling, code specific to the bytecode interpreter */
-
-#include <signal.h>
-#include <errno.h>
-#include "caml/config.h"
-#include "caml/memory.h"
-#include "caml/fail.h"
-#include "caml/finalise.h"
-#include "caml/osdeps.h"
-#include "caml/signals.h"
-#include "caml/signals_machdep.h"
-
-#ifdef _WIN32
-typedef void (*sighandler)(int sig);
-extern sighandler caml_win32_signal(int sig, sighandler action);
-#define signal(sig,act) caml_win32_signal(sig,act)
-#endif
-
-static void handle_signal(int signal_number)
-{
-  int saved_errno;
-  /* Save the value of errno (PR#5982). */
-  saved_errno = errno;
-#if !defined(POSIX_SIGNALS) && !defined(BSD_SIGNALS)
-  signal(signal_number, handle_signal);
-#endif
-  caml_record_signal(signal_number);
-  errno = saved_errno;
-}
-
-int caml_set_signal_action(int signo, int action)
-{
-  void (*act)(int signo), (*oldact)(int signo);
-#ifdef POSIX_SIGNALS
-  struct sigaction sigact, oldsigact;
-#endif
-
-  switch (action) {
-  case 0:  act = SIG_DFL; break;
-  case 1:  act = SIG_IGN; break;
-  default: act = handle_signal; break;
-  }
-
-#ifdef POSIX_SIGNALS
-  sigact.sa_handler = act;
-  sigemptyset(&sigact.sa_mask);
-  sigact.sa_flags = 0;
-  if (sigaction(signo, &sigact, &oldsigact) == -1) return -1;
-  oldact = oldsigact.sa_handler;
-#else
-  oldact = signal(signo, act);
-  if (oldact == SIG_ERR) return -1;
-#endif
-  if (oldact == handle_signal)
-    return 2;
-  else if (oldact == SIG_IGN)
-    return 1;
-  else
-    return 0;
-}
